--- qmat/playground/diff_eqs/__init__.py
@@ -1,10 +1,0 @@
-
-from qmat.playground.diff_eqs.de_solver import DESolver
-from qmat.playground.diff_eqs.burgers import Burgers
-from qmat.playground.diff_eqs.dahlquist2 import Dahlquist2
-
-__all__ = [
-    "DESolver",
-    "Burgers",
-    "Dahlquist2",
-]
