+++ qmat/playgrounds/martin/diff_eqs/__init__.py
@@ -1,0 +1,10 @@
+
+from qmat.playgrounds.martin.diff_eqs.de_solver import DESolver
+from qmat.playgrounds.martin.diff_eqs.burgers import Burgers
+from qmat.playgrounds.martin.diff_eqs.dahlquist2 import Dahlquist2
+
+__all__ = [
+    "DESolver",
+    "Burgers",
+    "Dahlquist2",
+]
