+++ qmat/playgrounds/martin/diff_eqs/de_solver.py
