+++ qmat/playgrounds/martin/ex_burgers_plot.py
