+++ qmat/playgrounds/martin/tests/test_dahlquist.py
@@ -1,16 +1,16 @@
 import numpy as np
-from qmat.playground.diff_eqs.two_freq import TwoFreq
+from qmat.playgrounds.martin.diff_eqs.dahlquist import Dahlquist
 from time_integration.sdc_integration import SDCIntegration
-from qmat.playground.time_integration.rk_integration import RKIntegration
+from qmat.playgrounds.martin.time_integration.rk_integration import RKIntegration
 
 
-def test_dahlquist2():
+def test_dahlquist():
+    u0 = np.array([1.0])  # Initial condition
     T: float = 4 * np.pi  # Time interval
     T: float = 0.5  # Time interval
     t: float = 0.0  # Starting time
 
-    two_freq: TwoFreq = TwoFreq(lam1=1.0j, lam2=0.1j)
-    u0 = two_freq.initial_u0()
+    dahlquist: Dahlquist = Dahlquist(lam1=1.0j, lam2=0.1j)
 
     for time_integration in ["rk1", "rk2", "rk4", "sdc"]:
         print("="*80)
@@ -18,7 +18,7 @@ def test_dahlquist2():
         print("="*80)
         results = []
 
-        u_analytical = two_freq.u_solution(u0, t=T)
+        u_analytical = dahlquist.u_solution(u0, t=T)
 
         for nt in range(4):
 
@@ -27,18 +27,18 @@ def test_dahlquist2():
 
             dt = T / num_timesteps
 
-            u0 = two_freq.initial_u0()
+            u0 = dahlquist.initial_u0()
 
             u = u0.copy()
 
             if time_integration in RKIntegration.supported_methods:
                 rki = RKIntegration(method=time_integration)
 
-                u = rki.integrate_n(u, t, dt, num_timesteps, two_freq)
+                u = rki.integrate_n(u, t, dt, num_timesteps, dahlquist)
 
             elif time_integration == "sdc":
-                sdci = SDCIntegration(num_nodes=3, node_type="LEGENDRE", quad_type="LOBATTO", num_sweeps=2)
-                u = sdci.integrate_n(u, t, dt, num_timesteps, two_freq)
+                sdci = SDCIntegration(num_nodes=3, node_type="LEGENDRE", quad_type="LOBATTO", num_sweeps=1)
+                u = sdci.integrate_n(u, t, dt, num_timesteps, dahlquist)
 
             else:
                 raise Exception("TODO")
@@ -62,13 +62,13 @@ def test_dahlquist2():
             assert np.abs(results[-1]["conv"]-1.0) < 1e-2
 
         elif time_integration == "rk2":
-            assert results[-1]["error"] < 1e-5
+            assert results[-1]["error"] < 1e-4
             assert np.abs(results[-1]["conv"]-2.0) < 1e-3
 
         elif time_integration == "rk4":
-            assert results[-1]["error"] < 1e-11
-            assert np.abs(results[-1]["conv"]-4.0) < 1e-2
+            assert results[-1]["error"] < 1e-9
+            assert np.abs(results[-1]["conv"]-4.0) < 1e-4
 
         elif time_integration == "sdc":
-            assert results[-1]["error"] < 1e-5
-            assert np.abs(results[-1]["conv"]-3.0) < 1e-3
+            assert results[-1]["error"] < 1e-4
+            assert np.abs(results[-1]["conv"]-2.0) < 1e-3
