+++ qmat/playgrounds/martin/tests/test_dahlquist2.py
@@ -1,7 +1,7 @@
 import numpy as np
-from qmat.playground.diff_eqs.dahlquist2 import Dahlquist2
+from qmat.playgrounds.martin.diff_eqs.dahlquist2 import Dahlquist2
 from time_integration.sdc_integration import SDCIntegration
-from qmat.playground.time_integration.rk_integration import RKIntegration
+from qmat.playgrounds.martin.time_integration.rk_integration import RKIntegration
 
 
 def test_dahlquist2():
