+++ qmat/playgrounds/martin/time_integration/rk_integration.py
@@ -1,6 +1,6 @@
 from typing import List
 import numpy as np
-from qmat.playground.diff_eqs.de_solver import DESolver
+from qmat.playgrounds.martin.diff_eqs.de_solver import DESolver
 
 
 class RKIntegration:
