+++ qmat/playgrounds/martin/time_integration/sdc_integration.py
@@ -1,5 +1,5 @@
 import numpy as np
-from qmat.playground.diff_eqs.de_solver import DESolver
+from qmat.playgrounds.martin.diff_eqs.de_solver import DESolver
 
 
 class SDCIntegration:
